scripts/requirements.txt
yfinance==0.2.37
psycopg2-binary==2.9.9
